refactor(source): move source to its own package [WIP]
